refactor: split options
